+++ dcicutils/sheet_utils.py
@@ -1,10 +1,16 @@
+import chardet
 import copy
+import csv
+import io
+import openpyxl
 
 from dcicutils.common import AnyJsonData
-from openpyxl import load_workbook
+from dcicutils.lang_utils import conjoined_list, maybe_pluralize
+from dcicutils.misc_utils import ignored
 from openpyxl.worksheet.worksheet import Worksheet
 from openpyxl.workbook.workbook import Workbook
-from typing import Any, Dict, List, Optional, Union
+from tempfile import TemporaryFile
+from typing import Any, Dict, Iterable, List, Optional, Union
 
 
 Header = str
@@ -12,6 +18,51 @@ Headers = List[str]
 ParsedHeader = List[Union[str, int]]
 ParsedHeaders = List[ParsedHeader]
 SheetCellValue = Union[int, float, str]
+SheetRow = List[SheetCellValue]
+CsvReader = type(csv.reader(TemporaryFile()))
+
+
+def unwanted_kwargs(*, context, kwargs, context_plural=False, detailed=False):
+    if kwargs:
+        unwanted = [f"{argname}={value!r}" if detailed else argname
+                    for argname, value in kwargs.items()
+                    if value is not None]
+        if unwanted:
+            does_not = "don't" if context_plural else "doesn't"
+            raise ValueError(f"{context} {does_not} use"
+                             f" {maybe_pluralize(unwanted, 'keyword argument')} {conjoined_list(unwanted)}.")
+
+
+def prefer_number(value: SheetCellValue):
+    if isinstance(value, str):  # the given value might be an int or float, in which case just fall through
+        if not value:
+            return None
+        value = value
+        ch0 = value[0]
+        if ch0 == '+' or ch0 == '-' or ch0.isdigit():
+            try:
+                return int(value)
+            except Exception:
+                pass
+            try:
+                return float(value)
+            except Exception:
+                pass
+        # If we couldn't parse it as an int or float, fall through to returning the original value
+        pass
+    return value
+
+
+def open_text_input_file_respecting_byte_order_mark(filename):
+    """
+    Opens a file for text input, respecting a byte-order mark (BOM).
+    """
+    with io.open(filename, 'rb') as fp:
+        leading_bytes = fp.read(4 * 8)  # 4 bytes is all we need
+        bom_info = chardet.detect(leading_bytes)
+        detected_encoding = bom_info and bom_info.get('encoding')  # tread lightly
+
+    return io.open(filename, 'r', encoding=detected_encoding)
 
 
 class ItemTools:
@@ -90,7 +141,7 @@ class ItemTools:
         return parent
 
     @classmethod
-    def parse_value(cls, value: SheetCellValue) -> AnyJsonData:
+    def parse_item_value(cls, value: SheetCellValue) -> AnyJsonData:
         if isinstance(value, str):
             lvalue = value.lower()
             # TODO: We could consult a schema to make this less heuristic, but this may do for now
@@ -101,19 +152,14 @@ class ItemTools:
             elif lvalue == 'null' or lvalue == '':
                 return None
             elif '|' in value:
-                return [cls.parse_value(subvalue) for subvalue in value.split('|')]
+                if value == '|':  # Use '|' for []
+                    return []
+                else:
+                    if value.endswith("|"):  # Use 'foo|' for ['foo']
+                        value = value[:-1]
+                    return [cls.parse_item_value(subvalue) for subvalue in value.split('|')]
             else:
-                ch0 = value[0]
-                if ch0 == '+' or ch0 == '-' or ch0.isdigit():
-                    try:
-                        return int(value)
-                    except Exception:
-                        pass
-                    try:
-                        return float(value)
-                    except Exception:
-                        pass
-                return value
+                return prefer_number(value)
         else:  # presumably a number (int or float)
             return value
 
@@ -128,24 +174,135 @@ class ItemTools:
             cls.set_path_value(datum[key], more_path, value)
 
 
-class WorkbookManager:
+# TODO: Consider whether this might want to be an abstract base class. Some change might be needed.
+#
+# Doug thinks we might want (metaclass=ABCMeta) here to make this an abstract base class.
+# I am less certain but open to discussion. Among other things, as implemented now,
+# the __init__ method here needs to run and the documentation says that ABC's won't appear
+# in the method resolution order. -kmp 17-Aug-2023
+# See also discussion at https://github.com/4dn-dcic/utils/pull/276#discussion_r1297775535
+class AbstractTableSetManager:
+    """
+    The TableSetManager is the spanning class of anything that wants to be able to load a table set,
+    regardless of what it wants to load it from. To do this, it must support a load method
+    that takes a filename and returns the file content in the form:
+        {
+            "Sheet1": [
+                          {...representation of row1 as some kind of dict...},
+                          {...representation of row2 as some kind of dict...}
+                      ],
+            "Sheet2": [...],
+            ...,
+        }
+    Note that at this level of abstraction, we take no position on what form of representation is used
+    for the rows, as long as it is JSON data of some kind. It might be
+         {"col1": "val1", "col2": "val2", ...}
+    or it might be something more structured like
+         {"something": "val1", {"something_else": ["val2"]}}
+    Additionally, the values stored might be altered as well. In particular, the most likely alteration
+    is to turn "123" to 123 or "" to None, though the specifics of whether and how such transformations
+    happen is not constrained by this class.
+    """
+
+    def __init__(self, **kwargs):
+        unwanted_kwargs(context=self.__class__.__name__, kwargs=kwargs)
+
+    # TODO: Consider whether this should be an abstractmethod (but first see detailed design note at top of class.)
+    @classmethod
+    def load(cls, filename: str) -> Dict[str, List[AnyJsonData]]:
+        """
+        Reads a filename and returns a dictionary that maps sheet names to rows of dictionary data.
+        For more information, see documentation of AbstractTableSetManager.
+        """
+        raise NotImplementedError(f".load(...) is not implemented for {cls.__name__}.")
+
+
+class BasicTableSetManager(AbstractTableSetManager):
+    """
+    A BasicTableManager provides some structure that most kinds of parsers will need.
+    In particular, everything will likely need some way of storing headers and some way of storing content
+    of each sheet. Even a csv file, which doesn't have multiple tabs can be seen as the degenerate case
+    of this where there's only one set of headers and only one block of content.
+    """
+
+    def __init__(self, filename: str, **kwargs):
+        super().__init__(**kwargs)
+        self.filename: str = filename
+        self.headers_by_tabname: Dict[str, Headers] = {}
+        self.content_by_tabname: Dict[str, List[AnyJsonData]] = {}
+        self.reader_agent: Any = self._get_reader_agent()
+
+    def tab_headers(self, tabname: str) -> Headers:
+        return self.headers_by_tabname[tabname]
+
+    def tab_content(self, tabname: str) -> List[AnyJsonData]:
+        return self.content_by_tabname[tabname]
+
+    def _create_tab_processor_state(self, tabname: str) -> Any:
+        """
+        This method provides for the possibility that some parsers will want auxiliary state,
+        (such as parsed headers or a line count or a table of temporary names for objects to cross-link
+        or some other such feature) that it carries with it as it moves from line to line parsing things.
+        Subclasses might therefore want to make this do something more interesting.
+        """
+        ignored(tabname)  # subclasses might need this, but we don't
+        return None
+
+    def _get_reader_agent(self) -> Any:
+        """This function is responsible for opening the workbook and returning a workbook object."""
+        raise NotImplementedError(f"._get_reader_agent() is not implemented for {self.__class__.__name__}.")
+
+    def load_content(self) -> Any:
+        raise NotImplementedError(f".load_content() is not implemented for {self.__class__.__name__}.")
+
+
+class TableSetManager(BasicTableSetManager):
 
     @classmethod
-    def load_workbook(cls, filename: str):
-        wb = cls(filename)
-        return wb.load_content()
+    def load(cls, filename: str) -> AnyJsonData:
+        table_set_manager: TableSetManager = cls(filename)
+        return table_set_manager.load_content()
 
-    def __init__(self, filename: str):
-        self.filename: str = filename
-        self.workbook: Optional[Workbook] = None
-        self.headers_by_sheetname: Dict[str, List[str]] = {}
-        self.content_by_sheetname: Dict[str, List[Any]] = {}
+    def __init__(self, filename: str, **kwargs):
+        super().__init__(filename=filename, **kwargs)
 
-    def sheet_headers(self, sheetname: str) -> List[str]:
-        return self.headers_by_sheetname[sheetname]
+    @property
+    def tabnames(self) -> List[str]:
+        raise NotImplementedError(f".tabnames is not implemented for {self.__class__.__name__}..")
 
-    def sheet_content(self, sheetname: str) -> List[Any]:
-        return self.content_by_sheetname[sheetname]
+    def _raw_row_generator_for_tabname(self, tabname: str) -> Iterable[SheetRow]:
+        """
+        Given a tabname and a state (returned by _sheet_loader_state), return a generator for a set of row values.
+        """
+        raise NotImplementedError(f"._rows_for_tabname(...) is not implemented for {self.__class__.__name__}.")
+
+    def _process_row(self, tabname: str, state: Any, row: List[SheetCellValue]) -> AnyJsonData:
+        """
+        This needs to take a state and whatever represents a row and
+        must return a list of objects representing column values.
+        What constitutes a processed up to the class, but other than that the result must be a JSON dictionary.
+        """
+        raise NotImplementedError(f"._process_row(...) is not implemented for {self.__class__.__name__}.")
+
+    def load_content(self) -> AnyJsonData:
+        for tabname in self.tabnames:
+            sheet_content = []
+            state = self._create_tab_processor_state(tabname)
+            for row_data in self._raw_row_generator_for_tabname(tabname):
+                processed_row_data: AnyJsonData = self._process_row(tabname, state, row_data)
+                sheet_content.append(processed_row_data)
+            self.content_by_tabname[tabname] = sheet_content
+        return self.content_by_tabname
+
+    @classmethod
+    def parse_cell_value(cls, value: SheetCellValue) -> AnyJsonData:
+        return prefer_number(value)
+
+
+class XlsxManager(TableSetManager):
+    """
+    This implements the mechanism to get a series of rows out of the sheets in an XLSX file.
+    """
 
     @classmethod
     def _all_rows(cls, sheet: Worksheet):
@@ -159,60 +316,182 @@ class WorkbookManager:
         for col in range(1, col_max + 1):
             yield col
 
-    def _load_headers(self, sheet: Worksheet):
-        headers: List[str] = [str(sheet.cell(row=1, column=col).value)
-                              for col in self._all_cols(sheet)]
-        self.headers_by_sheetname[sheet.title] = headers
+    @property
+    def tabnames(self) -> List[str]:
+        return self.reader_agent.sheetnames
 
-    def _load_row(self, *, sheet: Worksheet, row: int):
-        headers = self.sheet_headers(sheet.title)
-        row_dict: Dict[str, Any] = {headers[col-1]: sheet.cell(row=row, column=col).value
-                                    for col in self._all_cols(sheet)}
-        return row_dict
+    def _get_reader_agent(self) -> Workbook:
+        return openpyxl.load_workbook(self.filename)
 
-    def load_content(self):
-        workbook: Workbook = load_workbook(self.filename)
-        self.workbook = workbook
-        for sheetname in workbook.sheetnames:
-            sheet: Worksheet = workbook[sheetname]
-            self._load_headers(sheet)
-            content = []
-            for row in self._all_rows(sheet):
-                row_dict = self._load_row(sheet=sheet, row=row)
-                content.append(row_dict)
-            self.content_by_sheetname[sheetname] = content
-        return self.content_by_sheetname
+    def _raw_row_generator_for_tabname(self, tabname: str) -> Iterable[SheetRow]:
+        sheet = self.reader_agent[tabname]
+        return (self._get_raw_row_content_tuple(sheet, row)
+                for row in self._all_rows(sheet))
+
+    def _get_raw_row_content_tuple(self, sheet: Worksheet, row: int) -> SheetRow:
+        return [sheet.cell(row=row, column=col).value
+                for col in self._all_cols(sheet)]
+
+    def _create_tab_processor_state(self, tabname: str) -> Headers:
+        sheet = self.reader_agent[tabname]
+        headers: Headers = [str(sheet.cell(row=1, column=col).value)
+                            for col in self._all_cols(sheet)]
+        self.headers_by_tabname[sheet.title] = headers
+        return headers
+
+    def _process_row(self, tabname: str, headers: Headers, row_data: SheetRow) -> AnyJsonData:
+        ignored(tabname)
+        return {headers[i]: self.parse_cell_value(row_datum)
+                for i, row_datum in enumerate(row_data)}
 
 
-class ItemManager(ItemTools, WorkbookManager):
+class ItemManagerMixin(BasicTableSetManager):
+    """
+    This can add functionality to a reader such as an XlsxManager or a CsvManager in order to make its rows
+    get handled like Items instead of just flat table rows.
+    """
 
-    def __init__(self, filename: str):
-        super().__init__(filename=filename)
-        self.patch_prototypes_by_sheetname: Dict[str, Dict] = {}
-        self.parsed_headers_by_sheetname: Dict[str, List[List[Union[int, str]]]] = {}
+    def __init__(self, filename: str, **kwargs):
+        super().__init__(filename=filename, **kwargs)
+        self.patch_prototypes_by_tabname: Dict[str, Dict] = {}
+        self.parsed_headers_by_tabname: Dict[str, ParsedHeaders] = {}
 
-    def sheet_patch_prototype(self, sheetname: str) -> Dict:
-        return self.patch_prototypes_by_sheetname[sheetname]
+    def sheet_patch_prototype(self, tabname: str) -> Dict:
+        return self.patch_prototypes_by_tabname[tabname]
 
-    def sheet_parsed_headers(self, sheetname: str) -> List[List[Union[int, str]]]:
-        return self.parsed_headers_by_sheetname[sheetname]
+    def sheet_parsed_headers(self, tabname: str) -> ParsedHeaders:
+        return self.parsed_headers_by_tabname[tabname]
 
-    def _load_headers(self, sheet: Worksheet):
-        super()._load_headers(sheet)
-        self._compile_sheet_headers(sheet.title)
+    def _create_tab_processor_state(self, tabname: str) -> ParsedHeaders:
+        super()._create_tab_processor_state(tabname)
+        # This will create state that allows us to efficiently assign values in the right place on each row
+        # by setting up a prototype we can copy and then drop values into.
+        self._compile_sheet_headers(tabname)
+        return self.sheet_parsed_headers(tabname)
 
-    def _compile_sheet_headers(self, sheetname: str):
-        headers = self.headers_by_sheetname[sheetname]
-        parsed_headers = self.parse_sheet_headers(headers)
-        self.parsed_headers_by_sheetname[sheetname] = parsed_headers
-        prototype = self.compute_patch_prototype(parsed_headers)
-        self.patch_prototypes_by_sheetname[sheetname] = prototype
+    def _compile_sheet_headers(self, tabname: str):
+        headers = self.headers_by_tabname[tabname]
+        parsed_headers = ItemTools.parse_sheet_headers(headers)
+        self.parsed_headers_by_tabname[tabname] = parsed_headers
+        prototype = ItemTools.compute_patch_prototype(parsed_headers)
+        self.patch_prototypes_by_tabname[tabname] = prototype
 
-    def _load_row(self, *, sheet: Worksheet, row: int):
-        parsed_headers = self.sheet_parsed_headers(sheet.title)
-        patch_item = copy.deepcopy(self.sheet_patch_prototype(sheet.title))
-        for col in self._all_cols(sheet):
-            value = sheet.cell(row=row, column=col).value
-            parsed_value = self.parse_value(value)
-            self.set_path_value(patch_item, parsed_headers[col - 1], parsed_value)
+    def _process_row(self, tabname: str, parsed_headers: ParsedHeaders, row_data: SheetRow) -> AnyJsonData:
+        patch_item = copy.deepcopy(self.sheet_patch_prototype(tabname))
+        for i, value in enumerate(row_data):
+            parsed_value = self.parse_cell_value(value)
+            ItemTools.set_path_value(patch_item, parsed_headers[i], parsed_value)
         return patch_item
+
+    @classmethod
+    def parse_cell_value(cls, value: SheetCellValue) -> AnyJsonData:
+        return ItemTools.parse_item_value(value)
+
+
+class ItemXlsxManager(ItemManagerMixin, XlsxManager):
+    """
+    This layers item-style row processing functionality on an XLSX file.
+    """
+    pass
+
+
+class CsvManager(TableSetManager):
+    """
+    This implements the mechanism to get a series of rows out of the sheet in a csv file,
+    returning a result that still looks like there could have been multiple tabs.
+    """
+
+    DEFAULT_TAB_NAME = 'Sheet1'
+
+    def __init__(self, filename: str, tab_name=None, **kwargs):
+        super().__init__(filename=filename, **kwargs)
+        self.tab_name = tab_name or self.DEFAULT_TAB_NAME
+
+    @property
+    def tabnames(self) -> List[str]:
+        return [self.tab_name]
+
+    def _get_reader_agent(self) -> CsvReader:
+        return self._get_csv_reader(self.filename)
+
+    @classmethod
+    def _get_csv_reader(cls, filename) -> CsvReader:
+        return csv.reader(open_text_input_file_respecting_byte_order_mark(filename))
+
+    PAD_TRAILING_TABS = True
+
+    def _raw_row_generator_for_tabname(self, tabname: str) -> Iterable[SheetRow]:
+        headers = self.tab_headers(tabname)
+        n_headers = len(headers)
+        for row_data in self.reader_agent:
+            n_cols = len(row_data)
+            if self.PAD_TRAILING_TABS and n_cols < n_headers:
+                row_data = row_data + [''] * (n_headers - n_cols)
+            yield row_data
+
+    def _create_tab_processor_state(self, tabname: str) -> Headers:
+        headers: Optional[Headers] = self.headers_by_tabname.get(tabname)
+        if headers is None:
+            self.headers_by_tabname[tabname] = headers = self.reader_agent.__next__()
+        return headers
+
+    def _process_row(self, tabname: str, headers: Headers, row_data: SheetRow) -> AnyJsonData:
+        ignored(tabname)
+        return {headers[i]: self.parse_cell_value(row_datum)
+                for i, row_datum in enumerate(row_data)}
+
+
+class ItemCsvManager(ItemManagerMixin, CsvManager):
+    """
+    This layers item-style row processing functionality on a CSV file.
+    """
+    pass
+
+
+class TsvManager(CsvManager):
+    """
+    TSV files are just CSV files with tabs instead of commas as separators.
+    (We do not presently handle any escaping of strange characters. May need to add handling for backslash escaping.)
+    """
+    @classmethod
+    def _get_csv_reader(cls, filename) -> CsvReader:
+        return csv.reader(open_text_input_file_respecting_byte_order_mark(filename), delimiter='\t')
+
+
+class ItemTsvManager(ItemManagerMixin, TsvManager):
+    """
+    This layers item-style row processing functionality on a TSV file.
+    """
+    pass
+
+
+class ItemManager(AbstractTableSetManager):
+    """
+    This class will open a .xlsx or .csv file and load its content in our standard format.
+    (See more detailed description in AbstractTableManager.)
+    """
+
+    @classmethod
+    def create_implementation_manager(cls, filename: str, **kwargs) -> BasicTableSetManager:
+        if filename.endswith(".xlsx"):
+            # unwanted_kwargs(context="ItemManager for .xlsx files", kwargs=kwargs)
+            reader_agent = ItemXlsxManager(filename, **kwargs)
+        elif filename.endswith(".csv"):
+            tab_name = kwargs.pop('tab_name', None)
+            # unwanted_kwargs(context="ItemManager for .csv files", kwargs=kwargs)
+            reader_agent = ItemCsvManager(filename, tab_name=tab_name, **kwargs)
+        elif filename.endswith(".tsv"):
+            tab_name = kwargs.pop('tab_name', None)
+            # unwanted_kwargs(context="ItemManager for .tsv files", kwargs=kwargs)
+            reader_agent = ItemTsvManager(filename, tab_name=tab_name, **kwargs)
+        else:
+            raise ValueError(f"Unknown file type: {filename}")
+        return reader_agent
+
+    @classmethod
+    def load(cls, filename: str, tab_name=None) -> AnyJsonData:
+        manager = cls.create_implementation_manager(filename, tab_name=tab_name)
+        return manager.load_content()
+
+
+load_items = ItemManager.load